requirements.txt
python-dotenv==1.0.1
faker==33.1.0
factory-boy==3.3.1